requirements.txt
replicate
Pillow